+++ BitBase/BitBase/Database.h
@@ -18,7 +18,7 @@ public:
     TickTableRead(const std::string& root_path, const std::string& exchange, const std::string& symbol);
     ~TickTableRead(void);
 
-    std::unique_ptr<Tick> get_tick(int tick_idx);
+    std::unique_ptr<Tick> get_tick(long long tick_idx);
     std::unique_ptr<Tick> get_next_tick(void);
 
 private:
@@ -37,13 +37,13 @@ public:
     static std::shared_ptr<Database> create(const std::string& root_path);
 
     const std::string get_attribute(const std::string& key, const std::string& default_string);
-    const int get_attribute(const std::string& key, int default_value);
+    const long long get_attribute(const std::string& key, long long default_value);
     const time_point_ms get_attribute(const std::string& key, const time_point_ms& default_date_time);
     const std::vector<std::string> get_attribute(const std::string& key, const std::vector<std::string>& default_string_vector);
     const std::unordered_set<std::string> get_attribute(const std::string& key, const std::unordered_set<std::string>& default_string_set);
 
     void set_attribute(const std::string& key, const std::string& string);
-    void set_attribute(const std::string& key, int value);
+    void set_attribute(const std::string& key, long long value);
     void set_attribute(const std::string& key, const time_point_ms& date_time);
     void set_attribute(const std::string& key, const std::vector<std::string>& string_vector);
     void set_attribute(const std::string& key, const std::unordered_set<std::string>& string_set);
@@ -56,7 +56,7 @@ public:
     std::unique_ptr<TickTableRead> open_tick_table_read(const std::string& exchange, const std::string& symbol);
 
     void extend_tick_data(const std::string& exchange, const std::string& symbol, uptrDatabaseTicks ticks, const time_point_ms& first_timestamp);
-    void extend_interval_data(const std::string& exchange, const std::string& symbol, const std::chrono::milliseconds interval, const Intervals& intervals_data, const time_point_ms& timestamp, int tick_idx);
+    void extend_interval_data(const std::string& exchange, const std::string& symbol, const std::chrono::milliseconds interval, const Intervals& intervals_data, const time_point_ms& timestamp, long long tick_idx);
 
     std::unique_ptr<Intervals> get_intervals(const std::string& exchange, const std::string& symbol, const time_point_ms& timestamp_start, const time_point_ms& timestamp_end, const std::chrono::milliseconds interval);
